moved graph saving funcs to utils module, added utils to tracking.
